Section Header formatting commands
Rolled in formatting commands from master branch of MLA13, but
modifying the styling to support centered section headers concisely.
The master branch of MLA13 defines section headers twice, first in the
clean manner suggested by kc9jud (and which I've also appropriated in
this fork), and secondly in a manner I suggested, with the second
manner centering the headers.